Create project_planning.md
As this project begins to grow, I wanted to separate out the planning and design documentation from the general README content. This file is meant to serve as a "home page" for aggregating all the planning, design, development, etc., resources in a single location. We can link to subpages as this project grows and we need more organization, but this should serve as a good starting point.